huge fix поправил время, больше не зависит от инициализации в приложении, почему зависело не ясно, проблема была в том, что использовался некорректный тип данных для хранения времени(uint32_t вместо uint16_t), что приводило к неверным расчётам в TimeDiff убрал закомментированный ненужный код поправил обновление ик-датчика, были перепутаны местами bits_count и word IRPKG теперь всегда устанавливается как 16 битное, так как это логично(дополнение ведущими нулями, а не следующими)

+++ compiler/library/IRPCB/1.0/source/yartos/yartos.h
@@ -19,7 +19,7 @@
 #if SYS_TIM_RESOLUTION == 32
 using systime_t = uint32_t;
 #elif SYS_TIM_RESOLUTION == 16
-using systime_t = uint32_t;
+using systime_t = uint16_t;
 #else
 #error "Bad SysTimer Resolution"
 #endif
@@ -307,17 +307,8 @@ void DbgCheckClassI();
 
 // Sys::Lock
 static inline systime_t GetSysTime() {
-    // return SYS_TIM->CNT;
     Sys::Lock();
     uint32_t time = Time_ + SYS_TIM->CNT;
-    // uint16_t currentCNT = SYS_TIM->CNT;
-    // if (lastCNT_ <= currentCNT) {
-    //     Time_ += currentCNT - lastCNT_;
-    // } else {
-    //     Time_ += 0x10000 - lastCNT_ + currentCNT;
-    // }
-    // lastCNT_ = currentCNT;
-    // systime_t time = Time_;
     Sys::Unlock();
     return time;
 }
